#298 Update documentation
