+++ development-requirements.txt
@@ -14,6 +14,7 @@ pytest-cov==6.2.1
 pytest-schema==0.1.2
 pytest==8.4.1
 senzing-core==0.4.1
+senzing-grpc==0.5.11
 twine==6.1.0
 virtualenv==20.32.0
 wheel==0.45.1
